[Refactor] (BSP): Rename header file from .h to .hpp
将头文件从 .h 更名为 .hpp，以符合 C++ 命名约定。
更新相关源文件的包含路径，确保编译正确。
此更改有助于提高代码的一致性和可读性。
