+++ Source/Core/main.cpp
@@ -5,7 +5,7 @@
 
 #include "FreeRTOS.h"
 #include "bsp_log.hpp"
-#include "bsp_uid.h"
+#include "bsp_uid.hpp"
 #include "chronolink.h"
 #include "harness.h"
 #include "mode_entry.h"
